__tests__/actions/context-groups.test.ts
import {
  createContextGroup,
  getContextGroups,
  getContextGroup,
  updateContextGroup,
  deleteContextGroup,
  addFileToContextGroup,
  removeFileFromContextGroup,
  getFilesInContextGroup
} from "@/actions/context-groups"
import { db } from "@/db/db"
import { contextGroupsTable, contextGroupFilesTable } from "@/db/schema/context-groups-schema"

// Mock the db module
jest.mock("@/db/db", () => ({
  insert: jest.fn(),
  query: {
    contextGroupsTable: {
      findMany: jest.fn(),
      findFirst: jest.fn()
    },
    contextGroupFilesTable: {
      findMany: jest.fn()
    }
  },
  update: jest.fn(),
  delete: jest.fn()
}))

// Mock the getUserId function
jest.mock("@/actions/auth/auth", () => ({
  getUserId: jest.fn(() => "test-user-id")
}))

describe("Context Groups Actions", () => {
  afterEach(() => {
    jest.clearAllMocks()
  })

  describe("createContextGroup", () => {
    it("should create a context group", async () => {
      const mockInsert = db.insert as jest.Mock
      mockInsert.mockResolvedValue([{ id: "new-group-id" }])

      const result = await createContextGroup({
        name: "Test Group",
        description: "Test Description",
        projectId: "project-id"
      })

      expect(result).toEqual({ id: "new-group-id" })
      expect(mockInsert).toHaveBeenCalledWith(contextGroupsTable)
    })
  })

  describe("getContextGroups", () => {
    it("should fetch context groups for a project", async () => {
      const mockFindMany = db.query.contextGroupsTable.findMany as jest.Mock
      mockFindMany.mockResolvedValue([{ id: "group-1" }, { id: "group-2" }])

      const result = await getContextGroups("project-id")

      expect(result).toEqual([{ id: "group-1" }, { id: "group-2" }])
      expect(mockFindMany).toHaveBeenCalled()
    })
  })

  // Add more test cases for other functions...

})